Move CGAffineTransform conformance to AnimatableProperty to a separate file
